requirements.txt
pypdf
langchain-community
langchain-core
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
torch
llama-cpp-python
boto3
pytest
